python/requirements.txt
Flask
paho-mqtt
gunicorn
wheel